+++ src/columns/maj_flt.rs
@@ -41,7 +41,7 @@ impl Column for MajFlt {
 #[cfg(target_os = "macos")]
 impl Column for MajFlt {
     fn add(&mut self, proc: &ProcessInfo) {
-        let raw_content = proc.curr_task.ptinfo.pti_pageins as u64;
+        let raw_content = proc.curr_task.pti_pageins as u64;
         let fmt_content = format!("{raw_content}");
 
         self.fmt_contents.insert(proc.pid, fmt_content);