+++ src/columns/usage_cpu.rs
@@ -50,10 +50,8 @@ impl Column for UsageCpu {
 #[cfg(target_os = "macos")]
 impl Column for UsageCpu {
     fn add(&mut self, proc: &ProcessInfo) {
-        let curr_time =
-            proc.curr_task.ptinfo.pti_total_user + proc.curr_task.ptinfo.pti_total_system;
-        let prev_time =
-            proc.prev_task.ptinfo.pti_total_user + proc.prev_task.ptinfo.pti_total_system;
+        let curr_time = proc.curr_task.pti_total_user + proc.curr_task.pti_total_system;
+        let prev_time = proc.prev_task.pti_total_user + proc.prev_task.pti_total_system;
         let usage_ms = (curr_time - prev_time) / 1000000u64;
         let interval_ms = proc.interval.as_secs() * 1000 + u64::from(proc.interval.subsec_millis());
         let usage = usage_ms as f64 * 100.0 / interval_ms as f64;